+++ R/gpdfit.R
@@ -6,16 +6,16 @@
 #' @keywords internal
 #' @export
 #' @param x a numeric vector. The sample from which to estimate the parameters.
-#' @return a list. Estimates for \eqn{k} and \eqn{\sigma}.
+#' @return a named list with components \code{k} and \code{sigma}.
 #'
 #' @details Here the parameter \eqn{k} is the negative of \eqn{k} in Zhang &
 #'   Stephens (2009).
 #'
 #' @note This function is primarily intended for internal use, but is exported
 #'   so that users can call it directly if desired. Users simply wishing to
-#'   compute LOO and WAIC should use the \code{\link{loo_and_waic}} function.
+#'   compute LOO should use the \code{\link{loo}} function.
 #'
-#' @seealso \code{\link{vgislw}}, \code{\link{loo_and_waic}},
+#' @seealso \code{\link{psislw}}, \code{\link{loo}},
 #' \code{\link{loo-package}}
 #'
 #' @references
@@ -25,15 +25,18 @@
 #'
 
 gpdfit <- function(x) {
-  n <- length(x)
+  N <- length(x)
   x <- sort.int(x, method = "quick")
   prior <- 3
-  m <- 80 + floor(sqrt(n))  # note: original paper used  20 + floor(sqrt(n))
-  b <- 1 / x[n] + (1 - sqrt(m / seq_min_half(m))) / prior / x[floor(n / 4 + 0.5)]
-  L <- vapply_seq(m, function(i) n * lx(b[i], x))
-  w <- 1 / vapply_seq(m, function(i) sum(exp(L - L[i])))
-  b <- sum(b * w)
-  k <- mean.default(log(1 - b * x))
-  sigma <- -k / b
+  M <- 80 + floor(sqrt(N))  # note: original paper used  20 + floor(sqrt(n))
+  mseq <- seq_len(M)
+  sM <- 1 - sqrt(M / (mseq - 0.5))
+  Nflr <- floor(N / 4 + 0.5)
+  b <- 1 / x[N] + sM / prior / x[Nflr]
+  l <- N * lx(b, x)
+  w <- 1 / vapply(mseq, FUN = function(j) sum(exp(l - l[j])), FUN.VALUE = 0)
+  bdotw <- sum(b * w)
+  k <- mean.default(log1p(-bdotw * x))
+  sigma <- -k / bdotw
   nlist(k, sigma)
 }