+++ tests/testthat/test_loo_and_waic.R
@@ -1,30 +1,39 @@
 library(loo)
 
-# test loo_and_waic -------------------------------------------------------
+# test loo and waic -------------------------------------------------------
+context("loo and waic")
+test_that("loo and waic return expected results", {
 
-context("loo_and_waic")
-set.seed(123)
-x <- matrix(rnorm(100), 20, 5)
+  set.seed(123)
+  x <- matrix(rnorm(5000), 100, 50)
+  ww <- waic(x)
+  wnms <- names(ww)
+  ll <- loo(x, cores = 1)
+  lnms <- names(ll)
+  waic_val <- unlist(ww[-grep("pointwise", wnms)])
+  loo_val <- unlist(ll[-grep("pointwise|pareto_k", lnms)])
 
-ans <- structure(list(elpd_loo = -1.24209890010025, p_loo = 3.68054227421264,
-                      elpd_waic = -1.74710271515681, p_waic = 4.1855460892692,
-                      looic = 2.4841978002005, waic = 3.49420543031363,
-                      se_elpd_loo = 0.511506733173098, se_p_loo = 0.255710175894716,
-                      se_elpd_waic = 0.526738752057132, se_p_waic = 0.30556903588444,
-                      se_looic = 1.0230134663462, se_waic = 1.05347750411426),
-                 .Names = c("elpd_loo", "p_loo", "elpd_waic", "p_waic", "looic",
-                            "waic", "se_elpd_loo", "se_p_loo", "se_elpd_waic",
-                            "se_p_waic", "se_looic", "se_waic"))
+  waic_ans <- structure(c(-25.1291601624985, 49.5591373599165, 50.258320324997,
+                          0.742909828603039, 1.05244649202819, 1.48581965720608),
+                        .Names = c("elpd_waic", "p_waic", "waic", "se_elpd_waic",
+                                   "se_p_waic", "se_waic"))
+  loo_ans <- structure(c(-24.2339828660691, 48.6639600634871, 48.4679657321382,
+                         0.706772789390649, 0.990389359906155, 1.4135455787813),
+                       .Names = c("elpd_loo", "p_loo", "looic", "se_elpd_loo",
+                                  "se_p_loo", "se_looic"))
 
-test_that("loo_and_waic returns correct result", {
-  Sys.setenv("R_TESTS" = "")
-  loo <- loo_and_waic(x, cores = 1)
-  val <- loo[-grep("pointwise|pareto_k", names(loo))]
-  expect_equal(val, ans)
 
-  expect_error(loo_and_waic(1:10), 'log_lik should be a matrix')
+  pareto_k <- ll$pareto_k
+  pareto_k_val <- list(mean = mean(pareto_k), range = range(pareto_k))
+  pareto_k_ans <- list(mean = 0.273775211941035,
+                       range = c(-0.245572962408798, 0.859988648019406))
 
-  loo_diff <- unlist(loo_and_waic_diff(loo, loo))
-  expect_true(all(loo_diff == 0))
+  expect_equivalent(loo_val, loo_ans)
+  expect_equivalent(waic_val, waic_ans)
+  vec <- 1:10
+  arr <- array(1:100, dim = c(2,5,10))
+  expect_error(loo(vec), 'log_lik should be a matrix')
+  expect_error(waic(arr), 'log_lik should be a matrix')
+
+  expect_equivalent(pareto_k_val, pareto_k_ans)
 })
-