+++ NEWS.md
@@ -1,2 +1,8 @@
-# loo 0.1
+# loo 0.1.0.9000
+* Better handling of special cases
+* Deprecates `loo_and_waic` function in favor of separate functions `loo` and
+`waic`
+* Deprecates `loo_and_waic_diff`. Use `compare` instead. 
+
+# loo 0.1.0
 * Initial release